requirements.txt
lxml==4.6.3
matplotlib==3.3.4
numpy==1.19.5
tqdm==4.62.2
torch==1.8.1
torchvision==0.9.1
pycocotools==2.0.2
Pillow==8.3.1
cython==0.29.24
